Refactor util method for rendering changes
