+++ mythx_cli/formatter/util.py
@@ -1,14 +1,9 @@
 """Utility functions for handling API requests and responses."""
 
-from collections import defaultdict
-from typing import List, Optional, Tuple, Union
+from typing import List, Union
 
 import click
-from mythx_models.response import (
-    AnalysisInputResponse,
-    DetectedIssuesResponse,
-    Severity,
-)
+from mythx_models.response import DetectedIssuesResponse, Severity
 
 SEVERITY_ORDER = (
     Severity.UNKNOWN,
@@ -17,71 +12,6 @@ SEVERITY_ORDER = (
     Severity.MEDIUM,
     Severity.HIGH,
 )
-
-
-def index_by_filename(
-    issues_list: List[Tuple[DetectedIssuesResponse, Optional[AnalysisInputResponse]]]
-):
-    """Index the given report/input responses by filename.
-
-    This will return a simplified, unified representation of the report/input payloads
-    returned by the MythX API. It is a mapping from filename to an iterable of issue
-    objects, which contain the report UUID, SWC ID, SWC title, short and long
-    description, severity, as well as the issue's line location in the source code.
-
-    This representation is meant to be passed on to the respective formatter, which
-    them visualizes the data.
-
-    :param issues_list: A list of two-tuples containing report and input responses
-    :return: A simplified mapping indexing issues by their file path
-    """
-    file_to_issues = defaultdict(list)
-
-    for resp, inp in issues_list:
-        for report in resp.issue_reports:
-            for issue in report.issues:
-                issue_entry = {
-                    "uuid": resp.uuid,
-                    "swcID": issue.swc_id,
-                    "swcTitle": issue.swc_title,
-                    "description": {
-                        "head": issue.description_short,
-                        "tail": issue.description_long,
-                    },
-                    "severity": issue.severity,
-                }
-
-                if issue.swc_id == "" or issue.swc_title == "" or not issue.locations:
-                    # skip issues with missing SWC or location data
-                    continue
-
-                source_formats = [loc.source_format for loc in issue.locations]
-                for loc in issue.locations:
-                    if loc.source_format != "text" and "text" in source_formats:
-                        # skip non-text locations when we have one attached to the issue
-                        continue
-
-                    for c in loc.source_map.components:
-                        source_list = loc.source_list or report.source_list
-                        if not (source_list and 0 <= c.file_id < len(source_list)):
-                            # skip issues whose srcmap file ID if out of range of the source list
-                            continue
-                        filename = source_list[c.file_id]
-
-                        if not inp.sources or filename not in inp.sources:
-                            # skip issues that can't be decoded to source location
-                            continue
-
-                        line = get_source_location_by_offset(
-                            inp.sources[filename]["source"], c.offset
-                        )
-                        issue_entry["line"] = line
-                        issue_entry["snippet"] = inp.sources[filename]["source"].split(
-                            "\n"
-                        )[line - 1]
-                        file_to_issues[filename].append(issue_entry)
-
-    return file_to_issues
 
 
 def get_source_location_by_offset(source: str, offset: int) -> int:
